+++ demo/src/main/java/com/example/demo/services/ProductService.java
@@ -4,8 +4,11 @@ import com.example.demo.dto.CreateProductDto;
 import com.example.demo.dto.GetProductsDto;
 import com.example.demo.dto.ProductPositionDto;
 import com.example.demo.dto.UpdateProductDto;
+import com.example.demo.exceptions.ProductIsInUserCartException;
 import com.example.demo.models.Category;
 import com.example.demo.models.Product;
+import com.example.demo.repositories.CartItemRepository;
+import com.example.demo.repositories.CategoryRepository;
 import com.example.demo.repositories.ProductRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
@@ -13,16 +16,19 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class ProductService {
     private final ProductRepository productRepository;
-    private final CategoryService categoryService;
+    private final CategoryRepository categoryRepository;
+    private final CartItemRepository cartItemRepository;
 
-    public ProductService(ProductRepository productRepository, CategoryService categoryService) {
+    public ProductService(ProductRepository productRepository, CategoryRepository categoryRepository, CartItemRepository cartItemRepository) {
         this.productRepository = productRepository;
-        this.categoryService = categoryService;
+        this.categoryRepository = categoryRepository;
+        this.cartItemRepository = cartItemRepository;
     }
 
     public List<GetProductsDto> getAllProducts() {
@@ -45,6 +51,13 @@ public class ProductService {
     }
 
     public void deleteProductById(Long id) {
+
+        boolean isAddedToCart = cartItemRepository.existsByProductId(id);
+
+        if(isAddedToCart) {
+            throw new ProductIsInUserCartException();
+        }
+
         productRepository.deleteById(id);
     }
     @Transactional
@@ -52,7 +65,8 @@ public class ProductService {
         Product product = productRepository.findById(updateProductDto.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + updateProductDto.getId()));
 
-            Category category = categoryService.getCategoryById(updateProductDto.getCategoryId());
+            Category category = categoryRepository.findById(updateProductDto.getCategoryId()).
+                    orElseThrow(() -> new EntityNotFoundException("Category not found"));
             product.setCategory(category);
             product.setName(updateProductDto.getName());
             product.setDescription(updateProductDto.getDescription());
@@ -69,7 +83,9 @@ public class ProductService {
         product.setPrice(createProductDto.getPrice());
         product.setDescription(createProductDto.getDescription());
         product.setPosition(newPosition);
-        product.setCategory(categoryService.getCategoryById(createProductDto.getCategoryId()));
+        Category cat = categoryRepository.findById(createProductDto.getCategoryId()).
+                orElseThrow(() -> new EntityNotFoundException("Category not found"));
+        product.setCategory(cat);
 
         return productRepository.save(product);
     }
